Small fixes

+++ src/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mhaan <mhaan@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/27 16:09:56 by mhaan         #+#    #+#                 */
-/*   Updated: 2022/11/15 11:52:10 by mhaan         ########   odam.nl         */
+/*   Updated: 2022/11/23 11:27:03 by mhaan         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 // prototypes
 int	ft_printf(const char *fmt, ...);
 int	fmt_switch(char fmt, va_list arg, int *len);
-int	put_uns_base(size_t n, int base, char fmt);
+int	put_uns_base(unsigned long long n, int base, char fmt);
 int	put_sign_dec(int n, char fmt);
 int	put_str_printf(char *str);
 int	put_chr_printf(char c);
